Rename terrorist package to terroristen (faction name)
